Updated client/src/App.js with useContext hook
Added useContext hook to the client/src/App.js file. The useContext hook was used to manage the user login state by passing the user into the client/src/App.js file. The user is checked before the login process is handled. If the user is Authenticated, the loging process kicks off and subsequently completes successfully and the user is allowed access into the system else the user is redirected to the User registration page